rtf_cli

Module commands

The various user facing commands exposed through the CLI.

We follow the git CLI design of splitting user facing commands into low level “plumbing” and high level “porcelain” categories.

Modules

plumbing

Lower level commands for running individual pieces of functionality from the framework.

porcelain

High level user commands for running the most common workflows and functionality exposed through the rtf CLI.

Functions

get_context_and_check_outdir

load_and_resolve_test_plan_from_github

Handles loading a test plan from github and displaying user facing errors

load_and_resolve_test_plan_from_local

Handles loading a local test plan and displaying user facing errors